chore: refactor and modularize code
- add routes and middleware directories
- refactor index.js in route and middleware modules
